+++ R/break.R
@@ -33,11 +33,15 @@ broken <- function(model, new_observation, ...) {
 #'
 #' # works for interactions as well
 #' model <- lm(Sepal.Length ~ Petal.Width*Species, data = iris)
+#' summary(model)
+#'
 #' new_observation <- iris[1,]
 #' br <- broken(model, new_observation)
+#' br
 #' plot(br)
 #'
 #' br2 <- broken(model, new_observation, predict.function = betas)
+#' br2
 #' plot(br2)
 
 broken.lm <- function(model, new_observation, ..., baseline = 0, predict.function = predict) {